requirements.txt
langchain
openai
python-dotenv
langchain_core
langchain_community